feat: add nack test
