--- src/solace_ai_connector/components/general/for_testing/give_nack_output/__init__.py
@@ -1,3 +1,0 @@
-"""Give nack output component"""
-
-from .give_nack_output import GiveNackOutput
