--- src/solace_ai_connector/components/general/for_testing/need_nack_input/__init__.py
@@ -1,3 +1,0 @@
-"""Need nack input component"""
-
-from .need_nack_input import NeedNackInput
